Skeleton project for icegrid.
This is my first project with the STM32F103. I got these little boards and
a programmer for $2 each off eBay. They feature an ARM Cortex M3, USB, and
a whole boatload of peripherals.

Took me a little while to get the tools up and running, but so far I'm
quite impressed. This project is a bare bones one that just blinks the
on-board LED. It uses our arm-none-minoca GCC toolchain.