Icegrid: Some initial Wifi connection server code.
This code (sometimes) exposes a wifi server that will collect credentials.
It currently does not read out the POST parameters, and seems to not
succeed on the first connection attempt.

Next up is to read the POST parameters, save the connection somewhere, and
attempt a client connection. Also figure out why requests seem to
sometimes fail.