Initial revision of the AudiDash firmware.
This firmware, which runs on the STM32F1 microcontrollers, controls a 1998
Audi A4 instrument cluster. It uses a "blue pill" STM32F1 board (super
cheap on eBay) and an ESP8266 Wifi board. It will beam out a network for
the first 60 seconds after being powered on, which you can connect to to
set the real WiFi credentials. These settings will be stored in the flash
for future power-ups, though there will still be a 60 second period in AP
mode where you can change these credentials at power up.

It then connects to the specified WiFi network using its saved
credentials, and listens on UDP port 8080 for incoming packets. The
packets are a simple text based format that allows for changing of the
indicator lights and gauges on the dashboard.

The example.py file shows how to turn on the left and right turn signals.
You'll need to adjust the IP address in that file to match the IP acquired
for your own network.